Rename index file, move non-vendor specific logic from clientcode into new index.js file, bump version
