+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fcp-client",
-  "version": "0.2",
+  "version": "2.0",
   "description": "Node interface for FCP API",
   "main": "index.js",
   "scripts": {
